+++ src/HSA/App.xaml.cs
@@ -116,7 +116,7 @@ public partial class App : Application
         sc.AddSingleton<WindowsUpdateClient>();
         sc.AddSingleton<DriverDownloader>();
         sc.AddSingleton<IDriverService, DriverService>();
-        sc.AddSingleton<IFirmwareService, FirmwareService>();
+        sc.AddSingleton<TestPageService>();
         sc.AddSingleton<IppConsumableSource>();
         sc.AddSingleton<WsdPrintConsumableSource>();
         sc.AddSingleton<PrinterEndpointDiscovery>();